+++ packages/solana/vue/src/composables/use-connection.ts
@@ -1,6 +1,7 @@
 import { createContext } from '../utils/context';
 import { Connection, clusterApiUrl } from '@solana/web3.js';
 import { computed, ComputedRef } from 'vue';
+import { ENDPOINTS, Endpoint } from '@axiajs/solana.core';
 
 export enum SolanaNetworks {
   mainnet = 'mainnet-beta',
@@ -8,10 +9,13 @@ export enum SolanaNetworks {
   testnet = 'testnet',
 }
 
-export type ConnectionContext = ComputedRef<Connection>;
+export type ConnectionContext = {
+  connection: ComputedRef<Connection>;
+  endpoint: ComputedRef<Endpoint>;
+};
 
 const [ConnectionContextProvider, useConnection, ConnectionInjectionKey] =
-  createContext<ComputedRef<Connection>>({
+  createContext<ConnectionContext>({
     name: 'ConnectionContext',
     strict: true,
     errorMessage: 'useConnection requires you to provide the connection hook',
@@ -22,8 +26,18 @@ export function initializeConnection(network: SolanaNetworks) {
     const _connection = new Connection(clusterApiUrl(network), 'recent');
 
     const connection = computed(() => _connection);
-    ConnectionContextProvider(connection);
-    return connection;
+    const endpoint = computed(
+      () => ENDPOINTS.find((e) => e.name === network) || ENDPOINTS[2] // Fallback to devnet
+    );
+
+    ConnectionContextProvider({
+      connection,
+      endpoint,
+    });
+    return {
+      connection,
+      endpoint,
+    };
   } catch (error) {
     console.error('There was a problem initializing connection', error);
   }