+++ packages/solana/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @axiajs/solana.core
 
+## 0.1.0-alpha.1
+
+### Patch Changes
+
+- Improve types
+- Updated dependencies
+  - @axiajs/solana.utils@0.1.0-alpha.1
+
 ## 0.1.0-alpha.0
 
 ### Minor Changes