+++ packages/solana/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @axiajs/solana.vue
 
+## 0.1.0-alpha.3
+
+### Patch Changes
+
+- Improve types
+- Updated dependencies
+  - @axiajs/solana.core@0.1.0-alpha.1
+
 ## 0.1.0-alpha.2
 
 ### Minor Changes